refactored xxe into core and creating patterns for other rules to follow
