+++ framework/codemodder-base/src/main/java/io/codemodder/remediation/xxe/XXEFixer.java
@@ -1,0 +1,10 @@
+package io.codemodder.remediation.xxe;
+
+import com.github.javaparser.ast.CompilationUnit;
+
+/** Interface for fixing XXEs. */
+interface XXEFixer {
+
+  /** A provider (for a given XML API) attempts to fix the given issue. */
+  <T> XXEFixAttempt tryFix(T issue, int line, Integer column, CompilationUnit cu);
+}
